refactor: move handler and message to Messenger namespace
