+++ EMS/core-bundle/src/Core/Messenger/Handler/JobHandler.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace EMS\CoreBundle\Core\MessageHandler;
+namespace EMS\CoreBundle\Core\Messenger\Handler;
 
-use EMS\CoreBundle\Core\Message\Job;
+use EMS\CoreBundle\Core\Messenger\Message\Job;
 use EMS\CoreBundle\Service\JobService;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
