+++ EMS/core-bundle/src/Core/Messenger/Message/Job.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EMS\CoreBundle\Core\Message;
+namespace EMS\CoreBundle\Core\Messenger\Message;
 
 readonly class Job implements AsyncMessageInterface
 {
