+++ elasticms-admin/config/packages/messenger.yaml
@@ -7,7 +7,7 @@ framework:
             failed: 'doctrine://default?queue_name=failed'
 
         routing:
-            'EMS\CoreBundle\Core\Message\AsyncMessageInterface': async
+            'EMS\CoreBundle\Core\Messenger\Message\AsyncMessageInterface': async
 
         buses:
             command.bus:
